chore(repo): refresh derived doc blocks
