+++ crates/convergio-graph/src/lib.rs
@@ -38,10 +38,12 @@ pub mod cluster;
 mod cluster_io;
 pub mod doc_link;
 pub mod drift;
+mod drift_git;
 pub mod error;
 pub mod meta;
 pub mod model;
 pub mod parse;
+mod parse_use;
 pub mod query;
 mod query_adrs;
 pub mod query_hints;
